+++ src/main/java/edu/ie3/datamodel/io/factory/timeseries/TimeSeriesMappingFactory.java
@@ -5,10 +5,10 @@
 */
 package edu.ie3.datamodel.io.factory.timeseries;
 
+import edu.ie3.datamodel.exceptions.FactoryException;
 import edu.ie3.datamodel.io.factory.EntityData;
 import edu.ie3.datamodel.io.factory.EntityFactory;
 import edu.ie3.datamodel.io.source.TimeSeriesMappingSource;
-import java.util.Collections;
 import java.util.List;
 import java.util.Set;
 import java.util.UUID;
@@ -17,6 +17,7 @@ import java.util.stream.Stream;
 
 public class TimeSeriesMappingFactory
     extends EntityFactory<TimeSeriesMappingSource.MappingEntry, EntityData> {
+  private static final String ENTITY = "entity";
   private static final String PARTICIPANT = "participant";
   private static final String TIME_SERIES = "timeSeries";
 
@@ -26,14 +27,21 @@ public class TimeSeriesMappingFactory
 
   @Override
   protected List<Set<String>> getFields(Class<?> entityClass) {
-    return Collections.singletonList(
+    return List.of(
+        Stream.of(ENTITY, TIME_SERIES).collect(Collectors.toSet()),
         Stream.of(PARTICIPANT, TIME_SERIES).collect(Collectors.toSet()));
   }
 
   @Override
   protected TimeSeriesMappingSource.MappingEntry buildModel(EntityData data) {
-    UUID participant = data.getUUID(PARTICIPANT);
     UUID timeSeries = data.getUUID(TIME_SERIES);
-    return new TimeSeriesMappingSource.MappingEntry(participant, timeSeries);
+
+    try {
+      UUID entity = data.getUUID(ENTITY);
+      return new TimeSeriesMappingSource.EntityMappingEntry(entity, timeSeries);
+    } catch (FactoryException e) {
+      UUID participant = data.getUUID(PARTICIPANT);
+      return new TimeSeriesMappingSource.ParticipantMappingEntry(participant, timeSeries);
+    }
   }
 }