+++ src/test/groovy/edu/ie3/datamodel/utils/validation/UniquenessValidationUtilsTest.groovy
@@ -12,6 +12,7 @@ import static edu.ie3.util.quantities.PowerSystemUnits.PU
 import static tech.units.indriya.unit.Units.METRE_PER_SECOND
 
 import edu.ie3.datamodel.exceptions.DuplicateEntitiesException
+import edu.ie3.datamodel.exceptions.ValidationException
 import edu.ie3.datamodel.io.source.TimeSeriesMappingSource
 import edu.ie3.datamodel.models.StandardUnits
 import edu.ie3.datamodel.models.input.AssetInput
@@ -23,6 +24,7 @@ import edu.ie3.datamodel.models.value.SolarIrradianceValue
 import edu.ie3.datamodel.models.value.TemperatureValue
 import edu.ie3.datamodel.models.value.WeatherValue
 import edu.ie3.datamodel.models.value.WindValue
+import edu.ie3.datamodel.utils.Try
 import edu.ie3.util.geo.GeoUtils
 import spock.lang.Specification
 import tech.units.indriya.quantity.Quantities
@@ -168,8 +170,8 @@ class UniquenessValidationUtilsTest extends Specification {
     given:
     UUID timeSeries = UUID.randomUUID()
     Set<TimeSeriesMappingSource.MappingEntry> uniqueEntries = [
-      new TimeSeriesMappingSource.MappingEntry(UUID.randomUUID(), timeSeries),
-      new TimeSeriesMappingSource.MappingEntry(UUID.randomUUID(), timeSeries),
+      new TimeSeriesMappingSource.ParticipantMappingEntry(UUID.randomUUID(), timeSeries),
+      new TimeSeriesMappingSource.ParticipantMappingEntry(UUID.randomUUID(), timeSeries),
     ]
 
     when:
@@ -183,29 +185,39 @@ class UniquenessValidationUtilsTest extends Specification {
     given:
     UUID participant = UUID.fromString("1f25eea2-20eb-4b6b-8f05-bdbb0e851e65")
 
-    Set<TimeSeriesMappingSource.MappingEntry> uniqueEntries = [
-      new TimeSeriesMappingSource.MappingEntry(participant, UUID.randomUUID()),
-      new TimeSeriesMappingSource.MappingEntry(participant, UUID.randomUUID()),
+    Set<TimeSeriesMappingSource.MappingEntry> uniqueParticipantEntries = [
+      new TimeSeriesMappingSource.ParticipantMappingEntry(participant, UUID.randomUUID()),
+      new TimeSeriesMappingSource.ParticipantMappingEntry(participant, UUID.randomUUID()),
+    ]
+
+    Set<TimeSeriesMappingSource.MappingEntry> uniqueEntityEntries = [
+      new TimeSeriesMappingSource.EntityMappingEntry(participant, UUID.randomUUID()),
+      new TimeSeriesMappingSource.EntityMappingEntry(participant, UUID.randomUUID()),
     ]
 
     when:
-    checkMappingEntryUniqueness(uniqueEntries)
+    def participantDuplicate = Try.ofVoid(() -> checkMappingEntryUniqueness(uniqueParticipantEntries), DuplicateEntitiesException)
+    def entityDuplicate = Try.ofVoid(() -> checkMappingEntryUniqueness(uniqueEntityEntries), DuplicateEntitiesException)
 
     then:
-    DuplicateEntitiesException de = thrown()
-    de.message == "'MappingEntry' entities with duplicated UUID key, but different field values found! " +
-        "Affected primary keys: [1f25eea2-20eb-4b6b-8f05-bdbb0e851e65]"
+    participantDuplicate.failure
+    participantDuplicate.exception.get().message == "'ParticipantMappingEntry' entities with duplicated UUID key, but different field values found! " +
+    "Affected primary keys: [1f25eea2-20eb-4b6b-8f05-bdbb0e851e65]"
+
+    entityDuplicate.failure
+    entityDuplicate.exception.get().message == "'EntityMappingEntry' entities with duplicated UUID key, but different field values found! " +
+    "Affected primary keys: [1f25eea2-20eb-4b6b-8f05-bdbb0e851e65]"
   }
 
   def "Checking if time based weather values are unique"() {
     given:
     ZonedDateTime time = ZonedDateTime.now()
     WeatherValue value = new WeatherValue(
-        GeoUtils.buildPoint(50d, 7d),
-        new SolarIrradianceValue(Quantities.getQuantity(10d, StandardUnits.SOLAR_IRRADIANCE), Quantities.getQuantity(10d, StandardUnits.SOLAR_IRRADIANCE)),
-        new TemperatureValue(Quantities.getQuantity(5d, Units.CELSIUS)),
-        new WindValue(Quantities.getQuantity(5d, DEGREE_GEOM), Quantities.getQuantity(10d, METRE_PER_SECOND))
-        )
+    GeoUtils.buildPoint(50d, 7d),
+    new SolarIrradianceValue(Quantities.getQuantity(10d, StandardUnits.SOLAR_IRRADIANCE), Quantities.getQuantity(10d, StandardUnits.SOLAR_IRRADIANCE)),
+    new TemperatureValue(Quantities.getQuantity(5d, Units.CELSIUS)),
+    new WindValue(Quantities.getQuantity(5d, DEGREE_GEOM), Quantities.getQuantity(10d, METRE_PER_SECOND))
+    )
 
     Set<TimeBasedValue<WeatherValue>> uniqueValues = [
       new TimeBasedValue<WeatherValue>(time, value),
@@ -223,11 +235,11 @@ class UniquenessValidationUtilsTest extends Specification {
     given:
     ZonedDateTime time = ZonedDateTime.now()
     WeatherValue value = new WeatherValue(
-        GeoUtils.buildPoint(50d, 7d),
-        new SolarIrradianceValue(Quantities.getQuantity(10d, StandardUnits.SOLAR_IRRADIANCE), Quantities.getQuantity(10d, StandardUnits.SOLAR_IRRADIANCE)),
-        new TemperatureValue(Quantities.getQuantity(5d, Units.CELSIUS)),
-        new WindValue(Quantities.getQuantity(5d, DEGREE_GEOM), Quantities.getQuantity(10d, METRE_PER_SECOND))
-        )
+    GeoUtils.buildPoint(50d, 7d),
+    new SolarIrradianceValue(Quantities.getQuantity(10d, StandardUnits.SOLAR_IRRADIANCE), Quantities.getQuantity(10d, StandardUnits.SOLAR_IRRADIANCE)),
+    new TemperatureValue(Quantities.getQuantity(5d, Units.CELSIUS)),
+    new WindValue(Quantities.getQuantity(5d, DEGREE_GEOM), Quantities.getQuantity(10d, METRE_PER_SECOND))
+    )
     Set<TimeBasedValue<WeatherValue>> notUniqueValues = [
       new TimeBasedValue<WeatherValue>(time, value),
       new TimeBasedValue<WeatherValue>(time, value)